+++ src/types/unikProperties.ts
@@ -10,3 +10,6 @@ export enum PioneerBadgeGrades {
 /* Pioneer badge grades for number of uniks on chain*/
 export const PIONEER_INNOVATOR = 1500;
 export const PIONEER_EARLY_ADOPTER = 150000;
+
+/* Special properties */
+export const VERIFIED_URL_KEY_PREFIX = "Verified/URL/";